backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Pillow==10.1.0
extcolors==1.0.0
requests==2.31.0
python-dotenv==1.0.0

# 数据库驱动（根据需要选择）
# SQLite (默认，无需额外安装)

# PostgreSQL (Vercel Postgres, Supabase, Neon 等)
psycopg2-binary==2.9.9

# MySQL (可选)
# mysql-connector-python==8.2.0
# PyMySQL==1.1.0

# Authentication
bcrypt==4.1.2
